+++ seqr/views/apis/anvil_workspace_api.py
@@ -2,6 +2,7 @@
 import json
 import time
 import tempfile
+from datetime import datetime
 
 from django.contrib.auth.decorators import user_passes_test
 from django.contrib.auth.views import redirect_to_login
@@ -11,6 +12,7 @@ from django.shortcuts import redirect
 from reference_data.models import GENOME_VERSION_LOOKUP
 from seqr.models import Project, CAN_EDIT
 from seqr.views.react_app import render_app_html
+from seqr.views.utils.dataset_utils import VCF_FILE_EXTENSIONS
 from seqr.views.utils.json_to_orm_utils import create_model_from_json
 from seqr.views.utils.json_utils import create_json_response
 from seqr.views.utils.file_utils import load_uploaded_file
@@ -18,7 +20,7 @@ from seqr.views.utils.terra_api_utils import add_service_account, has_service_ac
     TerraRefreshTokenFailedException
 from seqr.views.utils.pedigree_info_utils import parse_pedigree_table
 from seqr.views.utils.individual_utils import add_or_update_individuals_and_families
-from seqr.utils.communication_utils import safe_post_to_slack
+from seqr.utils.communication_utils import safe_post_to_slack, send_html_email
 from seqr.utils.file_utils import does_file_exist, file_iter, mv_file_to_gs
 from seqr.utils.logging_utils import SeqrLogger
 from seqr.views.utils.permissions_utils import is_anvil_authenticated, check_workspace_perm, login_and_policies_required
@@ -30,6 +32,12 @@ anvil_auth_required = user_passes_test(is_anvil_authenticated, login_url=GOOGLE_
 
 BLOCK_SIZE = 65536
 
+ANVIL_LOADING_EMAIL_DATE = '2021-12-13'
+ANVIL_LOADING_DELAY_EMAIL = """
+We have received your request to load data to seqr from AnVIL. Currently, the Broad Institute is holding an internal 
+retreat or closed for the winter break so we are unable to load data until the 2nd week of January 2022. We appreciate 
+your understanding and support of our research team taking some well-deserved time off and hope you also have a nice break.
+"""
 
 def get_vcf_samples(vcf_filename):
     byte_range = None if vcf_filename.endswith('.vcf') else (0, BLOCK_SIZE)
@@ -122,8 +130,8 @@ def create_project_from_workspace(request, namespace, name):
     # Validate the data path
     bucket_name = workspace_meta['workspace']['bucketName']
     data_path = 'gs://{bucket}/{path}'.format(bucket=bucket_name.rstrip('/'), path=request_json['dataPath'].lstrip('/'))
-    if not (data_path.endswith('.vcf') or data_path.endswith('.vcf.gz') or data_path.endswith('.vcf.bgz')):
-        error = 'Invalid VCF file format - file path must end with .vcf, .vcf.gz, or .vcf.bgz'
+    if not data_path.endswith(VCF_FILE_EXTENSIONS):
+        error = 'Invalid VCF file format - file path must end with {}'.format(' or '.join(VCF_FILE_EXTENSIONS))
         return create_json_response({'error': error}, status=400, reason=error)
     if not does_file_exist(data_path, user=request.user):
         error = 'Data file or path {} is not found.'.format(request_json['dataPath'])
@@ -172,6 +180,20 @@ def create_project_from_workspace(request, namespace, name):
 
     # Send a slack message to the slack channel
     _send_load_data_slack_msg(project, ids_path, data_path, request_json['sampleType'], request.user)
+
+    if ANVIL_LOADING_DELAY_EMAIL and ANVIL_LOADING_EMAIL_DATE and \
+            datetime.strptime(ANVIL_LOADING_EMAIL_DATE, '%Y-%m-%d') <= datetime.now():
+        try:
+            email_body = """Hi {user},
+            {email_content}
+            - The seqr team
+            """.format(
+                user=request.user.get_full_name() or request.user.email,
+                email_content=ANVIL_LOADING_DELAY_EMAIL,
+            )
+            send_html_email(email_body, subject='Delay in loading AnVIL in seqr', to=[request.user.email])
+        except Exception as e:
+            logger.error('AnVIL loading delay email error: {}'.format(e), request.user)
 
     return create_json_response({'projectGuid':  project.guid})